fix: SettingTableViewが表示されないバグの解決
